Basic implementation of Post filtering.
Actual filters comprised of two linked objects - FilterBase and Filter.
FlterBase stores info about the handler function, used for filtering, Filter
stores parameter for this function.
Result of handler function's invocation with Post and optional parameter is
stored as FilterResult object. FilterResults are cached in a lazy way, so if
at least one False result is cached for and-logic filter, there's no need to
apply any other filters, and they won't be applied, as soon as result is
apparent.
Filter objects are attached to Feed's in m2m fashion. Once it's done, every
single Post of the Feed should be a subject for filtering, not just the new
ones. Same goes for Filter detachment event.

What's missing:
 * Actual usage - there should be a filter for posts' selection upon display or
    export via feed. That should probably be done via manager object on a
    lowest-possible level, but that might need a lot of py-side processing,
    which should be avoided.
   Possible solution is to scrap all the work and make filtering pre-cached,
    so results would be calculated only for new posts and upon receiving m2m
    Feed-Filter relation or Filter changes.
 * Admin module description. All the tables should be exported, descriptions
    checked.
 * Signal-hook to drop FilterResult's upon Filter and FilterBase changes, at
    least. Might be extended to the whole pre-caching task.